Merge (12-06-2024)

+++ lib/models/cart_model.dart
@@ -1,6 +1,6 @@
 class Cart {
   final String itemName;
-  final double itemPrice;
+  final int itemPrice;
   final String itemImage;
   final String itemUnit;
   int qnt;
